MatrixPilot_fbw - reverted back again
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_fbw@2321 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/behavior.c
@@ -48,7 +48,7 @@ void init_behavior( void )
 }
 
 
-void setBehavior(int16_t newBehavior)
+void setBehavior(int newBehavior)
 {
 	desired_behavior.W = newBehavior ;
 	
@@ -195,7 +195,7 @@ void updateTriggerAction( void )
 		{
 			triggerActionSetValue( TRIGGER_ACTION == TRIGGER_PULSE_HIGH ) ;
 			
-			cyclesUntilStopTriggerAction = TRIGGER_PULSE_DURATION / (int32_t)25 ;
+			cyclesUntilStopTriggerAction = TRIGGER_PULSE_DURATION / (long)25 ;
 			cyclesUntilStartTriggerAction = 0 ;
 		}
 		else if ( TRIGGER_ACTION == TRIGGER_TOGGLE )
@@ -209,8 +209,8 @@ void updateTriggerAction( void )
 		{
 			triggerActionSetValue( TRIGGER_ACTION == TRIGGER_PULSE_HIGH ) ;
 			
-			cyclesUntilStopTriggerAction = TRIGGER_PULSE_DURATION / (int32_t)25 ;
-			cyclesUntilStartTriggerAction = TRIGGER_REPEAT_PERIOD / (int32_t)25 ;
+			cyclesUntilStopTriggerAction = TRIGGER_PULSE_DURATION / (long)25 ;
+			cyclesUntilStartTriggerAction = TRIGGER_REPEAT_PERIOD / (long)25 ;
 		}
 	}
 	else if ( cyclesUntilStartTriggerAction > 0 )
