+++ MatrixPilot/mode_switch.h
@@ -2,9 +2,9 @@
 #define FLIGHT_MODE_SWITCH_H
 
 
-inline int16_t flight_mode_switch_manual(void);
-inline int16_t flight_mode_switch_auto(void);
-inline int16_t flight_mode_switch_home(void);
+inline int flight_mode_switch_manual(void);
+inline int flight_mode_switch_auto(void);
+inline int flight_mode_switch_home(void);
 
 void flight_mode_switch_2pos_poll(void) ; // this is called at 40 hertz and scans the two postion switch option.
 void flight_mode_switch_check_set(void) ; // this is called at 2 hertz and changes the flight mode if req.
